refactor: Move the useMultiSelection hook to its Provider file
