+++ src/components/Home/HomeLayout.jsx
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types'
 import React, { useMemo, useState, useEffect } from 'react'
 import ContentFlexsearch from 'src/components/Home/ContentFlexsearch'
 import SearchHeader from 'src/components/Home/SearchHeader'
-import { useMultiSelection } from 'src/components/Hooks/useMultiSelection'
+import { useMultiSelection } from 'src/components/Contexts/MultiSelectionProvider'
 import { usePapersDefinitions } from 'src/components/Hooks/usePapersDefinitions'
 import FeaturedPlaceholdersList from 'src/components/Placeholders/FeaturedPlaceholdersList'
 import { useSearch } from 'src/components/Search/SearchProvider'
