+++ src/components/Multiselect/MultiselectPaperList.jsx
@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types'
 import React from 'react'
-import { useMultiSelection } from 'src/components/Hooks/useMultiSelection'
+import { useMultiSelection } from 'src/components/Contexts/MultiSelectionProvider'
 import Home from 'src/components/Views/Home'
 import PapersList from 'src/components/Views/PapersList'
