+++ src/components/Papers/PaperGroup.jsx
@@ -1,7 +1,7 @@
 import PropTypes from 'prop-types'
 import React from 'react'
 import { useNavigate } from 'react-router-dom'
-import { useMultiSelection } from 'src/components/Hooks/useMultiSelection'
+import { useMultiSelection } from 'src/components/Contexts/MultiSelectionProvider'
 import { CategoriesByPaper } from 'src/components/Papers/CategoriesByPaper'
 import KonnectorsCategories from 'src/components/Papers/KonnectorsCategories'
