+++ src/components/Papers/PaperItem.jsx
@@ -2,7 +2,7 @@ import cx from 'classnames'
 import PropTypes from 'prop-types'
 import React from 'react'
 import { useNavigate, useLocation } from 'react-router-dom'
-import { useMultiSelection } from 'src/components/Hooks/useMultiSelection'
+import { useMultiSelection } from 'src/components/Contexts/MultiSelectionProvider'
 import ExpirationAnnotation from 'src/components/Papers/ExpirationAnnotation'
 import { RemindersAnnotation } from 'src/components/Papers/RemindersAnnotation'
 import { generateReturnUrlToNotesIndex } from 'src/components/Papers/helpers'
